C2.0+C2.1: translator surface fixes + Ghidra pseudo-fn helpers
C2.0 — Translator stops inflating the Win32 import surface with
misclassifications:

* `stripWin32Casts(source)`: pre-strip well-known Win32 type casts
  (LPBOOL, HWND, LPSTR, etc.) before tree-sitter-c parses, since the
  parser can't tell typedefs from identifiers and was treating
  `(LPBOOL)&x` as a binary AND `(LPBOOL) & &x`.
* DAT_<addr>_<width> recognition in emitIdentifier and emitAssignment:
  Ghidra's typed-DAT notation (e.g. DAT_005f5500_2) now maps to
  heap.u16/u8 reads/writes.

Surface dropped 217 → 202 → 183 (Win32-only after pseudo-fn split).

C2.1 — runtime/ghidra-builtins.js implements the 19 pseudo-fns Ghidra
emits for x86 ops with no plain-C equivalent: CONCAT11/12/13/14/21/22/
24/31/44, CARRY1/2/4, SBORROW2/4, SCARRY2/4, LOCK, RtlUnwind, and
the placeholders Arguments + ExceptionList.

Translator now routes these to ../runtime/ghidra-builtins.js (separate
import line from win32.js) so the Win32 surface is genuinely
Win32-only. Pseudo-fns are pure helpers; no `heap` arg.

Validated by test/runtime/ghidra-builtins.test.js — 20 cases covering
all 19 helpers, with BigInt reference oracles for the carry/borrow
flags. fast-check exhaustively fuzzes input combinations.

102/102 tests pass.

+++ ported/auto/402027.js
@@ -36,7 +36,7 @@ export function FUN_00402027(heap) {
         local_c = local_b0;
         local_8 = heap.u32(0x005f1ff4) - local_b2;
         uVar4 = local_b0;
-        uVar3 = (uint)(local_b2 >>> 2);
+        uVar3 = (local_b2 >>> 2);
         puVar5 = heap.u32(0x005f1fec);
         do {
           for (; uVar3 != 0; uVar3 = uVar3 - 1) {
@@ -47,7 +47,7 @@ export function FUN_00402027(heap) {
           heap.setU32(__addr_local_b8, ((heap.u32(__addr_local_b8) + (local_a8 - local_b2))) >>> 0);
           puVar5 = (puVar5 + local_8);
           uVar4 = uVar4 - 1;
-          uVar3 = (uint)(local_b2 >>> 2);
+          uVar3 = (local_b2 >>> 2);
         } while (uVar4 != 0);
         FUN_00402aa4(heap);
         (heap.u32(heap.u32(0x005ebe60)))(__addr_local_b8);
