+++ ported/auto/41168c.js
@@ -7,6 +7,6 @@
 export function FUN_0041168c(heap) {
   let MVar1 = 0;
   MVar1 = mciSendStringA(0x005ec208, 0x0, 0, heap.u32(0x005e916c));
-  heap.setU32(0x005ec1c8, ((uint)(MVar1 == 0)) >>> 0);
+  heap.setU32(0x005ec1c8, ((MVar1 == 0)) >>> 0);
   return MVar1 == 0;
 }
