+++ ported/auto/4170f0.js
@@ -19,7 +19,7 @@ export function FUN_004170f0(heap, param_1) {
   }
   if ((heap.u32((param_1 + 0xc)) & 0x4000) != 0) {
     iVar1 = FUN_00418b60(heap, heap.u32((param_1 + 0x10)));
-    return -(uint)(iVar1 != 0);
+    return -(iVar1 != 0);
   }
   return 0;
 }
