+++ ported/auto/417860.js
@@ -40,7 +40,7 @@ export function FUN_00417860(heap, param_1, param_2, param_3, param_4, param_5, 
     if (param_5 == 0) {
       param_5 = heap.u32(0x005f0280);
     }
-    iVar2 = MultiByteToWideChar(heap, param_5, (-(uint)(param_7 != 0) & 8) + 1, param_2, param_3, 0x0, 0);
+    iVar2 = MultiByteToWideChar(heap, param_5, (-(param_7 != 0) & 8) + 1, param_2, param_3, 0x0, 0);
     if (iVar2 != 0) {
       lpWideCharStr = FUN_00413830(heap, 2, iVar2);
       if (lpWideCharStr != 0x0) {
