+++ ported/auto/42f6df.js
@@ -21,7 +21,7 @@ export function FUN_0042f6df(heap) {
   }
   heap.u32(heap.u32(0x005f88a8)) = in_AL;
   heap.u8(0x5f8d36) = heap.u32(0x005f8d36) + in_AL;
-  heap.setU32(0x005f8d36, (heap.u32(0x005f8d36) << 3 | (uint)((((heap.u32(0x005f8d36)) >>> 8) & 0xffffffff) >>> 0x15)) >>> 0);
+  heap.setU32(0x005f8d36, (heap.u32(0x005f8d36) << 3 | ((((heap.u32(0x005f8d36)) >>> 8) & 0xffffffff) >>> 0x15)) >>> 0);
   heap.setU32(0x005f88a8, (heap.u32(0x005f88a8) + 1) >>> 0);
   heap.setU32(0x005f88ac, (heap.u32(0x005f88ac) + 1) >>> 0);
   return;
