+++ ported/auto/4314ed.js
@@ -13,12 +13,12 @@ export function FUN_004314ed(heap) {
   do {
     uVar3 = uVar2 & 1;
     uVar2 = uVar2 >>> 1;
-    iVar1 = iVar1 + (uint)(uVar3 != 0);
+    iVar1 = iVar1 + (uVar3 != 0);
     uVar3 = heap.u32(0x0087c3e0);
   } while (uVar2 != 0);
   do {
     uVar2 = uVar3 >>> 1;
-    iVar1 = iVar1 + (uint)((uVar3 & 1) != 0);
+    iVar1 = iVar1 + ((uVar3 & 1) != 0);
     uVar3 = uVar2;
   } while (uVar2 != 0);
   return iVar1;
